refactor(event): content_sanitizer/media_service を分離 (#329 PR1)
libs.py 795行から以下を services/ に切り出した。

- services/content_sanitizer.py: ALLOWED_IFRAME_DOMAINS,
  _filter_iframe_attributes, _escape_unknown_html_tags
- services/media_service.py: ensure_pdf_thumbnail

libs.py 内の convert_markdown / apply_blog_output_to_event_detail は
新サービスから再 import して呼び出す。後方互換維持のための
re-export は行わず、import 元はすべて services パスに更新した。

#329 を 3PR に分割した最初の PR。残りは
PR2: content_generation + markdown_processor 分離（convert_markdown 分解含む）
PR3: libs.py 削除と残 import 整理。
